federico/requirements.txt
aiohttp==3.9.3
aiosignal==1.3.1
arabic-reshaper==3.0.0
asgiref==3.7.2
asn1crypto==1.5.1
attrs==23.2.0
certifi==2023.11.17
cffi==1.16.0
charset-normalizer==3.3.2
click==8.1.7
colorama==0.4.6
cryptography==41.0.5
cssselect2==0.7.0
distlib==0.3.7
Django==4.2.7
django-ckeditor==6.7.0
django-cors-headers==4.3.1
django-environ==0.11.2
django-filter==23.3
django-js-asset==2.1.0
django-storages==1.14.2
django-tables2==2.6.0
django-widget-tweaks==1.5.0
djangorestframework==3.14.0
filelock==3.13.1
freetype-py==2.3.0
frozenlist==1.4.1
html5lib==1.1
idna==3.4
lxml==4.9.3
multidict==6.0.5
mysql-connector-python==8.2.0
mysqlclient==2.2.1
numpy==1.26.1
oscrypto==1.3.0
pandas==2.2.0
Pillow==10.1.0
platformdirs==4.0.0
protobuf==4.21.12
psycopg2==2.9.9
pycairo==1.25.1
pycparser==2.21
pyHanko==0.20.1
pyhanko-certvalidator==0.24.1
pypdf==3.17.1
pypng==0.20220715.0
python-bidi==0.4.2
python-dateutil==2.8.2
pytz==2023.3.post1
PyYAML==6.0.1
qrcode==7.4.2
reportlab==4.0.7
requests==2.31.0
rlPyCairo==0.3.0
six==1.16.0
sqlparse==0.4.4
svglib==1.5.1
tinycss2==1.2.1
typing_extensions==4.8.0
tzdata==2023.3
tzlocal==5.2
uritools==4.0.2
urllib3==2.1.0
virtualenv==20.24.7
webencodings==0.5.1
xhtml2pdf==0.2.13
yarl==1.9.4
